refactor: déplace le composant LectureGuide
