+++ site/source/components/LectureGuide.tsx
