+++ site/source/components/RéductionDeCotisations/RépartitionValue.tsx
@@ -2,7 +2,7 @@ import { formatValue } from 'publicodes'
 import { useTranslation } from 'react-i18next'
 import { styled } from 'styled-components'
 
-import LectureGuide from '@/components/Simulation/LectureGuide'
+import LectureGuide from '@/components/LectureGuide'
 import { Grid } from '@/design-system/layout'
 import { SmallBody } from '@/design-system/typography/paragraphs'
